+++ database/seeders/DatabaseSeeder.php
@@ -190,6 +190,10 @@ class DatabaseSeeder extends Seeder
         ]);
 
 
+        $this->call([
+            CustomerSeeder::class,
+        ]);
+
 
 
     }