+++ resources/views/vm/index.blade.php
@@ -5,11 +5,13 @@
     @foreach ($vms as $vm)
         <x-card>
             <x-slot:head>
-                <x-show.header
-                    editUrl="{{ route('vm.edit', [$customer, $vm]) }}"
-                    deleteUrl="{{ route('vm.destroy', [$customer, $vm]) }}">
+                <x-show.header editUrl="{{ route('vm.edit', [$customer, $vm]) }}">
                     @if ($vm->remoteID AND $vm->remotePassword)
-                        <a href="rustdesk://connection/new/{{ $vm->remoteID }}?password={{ $vm->remotePassword }}" class=" bg-cerulean-500 text-gray-100 rounded-sm px-4 py-2 text-sm mr-5 hover:bg-cerulean-600">Verbinden</a>
+                        <x-input.linkbutton link="rustdesk://connection/new/{{ $vm->remoteID }}?password={{ $vm->remotePassword }}">
+                            <x-slot:label>
+                                <x-svg.software.rustdesk class="h-6 w-6 !fill-cerulean-500 hover:!fill-cerulean-400" />
+                            </x-slot:label>
+                        </x-input.linkbutton>
                     @endif
                     {{ $vm->name }}
                 </x-show.header>