+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [23-10-03]
+
+## Added
+
+- Global Remote Connection Search over Servers and VMs
+
+## Fixed
+
+- 1+n issue at UTM Search
+
+## Changed
+
+- Remote Connection Logo
+
+
+
 ## [23-10-02]
 
 ## Added
@@ -25,7 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - UTM External URL
 - Middelware isTechniker
 
-## Changes
+## Changed
 
 - Admin Bereich
 - Top Navigation bar
@@ -34,7 +50,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [23-09-27]
 
-## Changes
+## Changed
 
 - Admin Bereich
 
@@ -42,7 +58,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [23-09-26]
 
-## Changes
+## Changed
 
 - Desing Profile Page
 
@@ -57,7 +73,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Lizenzen Windows
 - Lizenzen Software
 
-## Changes
+## Changed
 
 - Desing
 
@@ -69,7 +85,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [23-09-09]
 
-## Changes
+## Changed
 
 - Desing
 - Update dependincies
@@ -82,7 +98,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Custom Font
 
-## Changes
+## Changed
 
 - Light Desing